+++ packages/http-client-csharp/generator/Microsoft.Generator.CSharp.ClientModel.TestProjects/Unbranded-TypeSpec/Unbranded-TypeSpec.tsp
@@ -204,6 +204,9 @@ model RoundTripModel {
 
   @doc("this is a model with required nullable properties")
   modelWithRequiredNullable: ModelWithRequiredNullableProperties;
+
+  @doc("Required bytes")
+  requiredBytes: bytes;
 }
 
 union DaysOfWeekExtensibleEnum {
